+++ src/map/map.hpp
@@ -7,8 +7,8 @@ Map is custom Canvas used for controlling all objects on the game map, such as T
 
 class Map : public Canvas {
 public:
-    Map(const Vector4f& body, const std::string& filePath);
-    Map(const Vector4f& body, int width, int height);
+    Map(const Vector4i& body, const std::string& filePath);
+    Map(const Vector4i& body, int width, int height);
     ~Map();
     Tile* GetTile(int x, int y) const;
 private:
